Add automark-actions-regex filter custom variable.
With this regex it simplifies to add whole set of commands to the filter
as an action. Emacs convection now uses a common prefix for all the
commands and variables in the same package/group so regexes like
"^groupstring-" may filter all the commands without needing to add them
one by one in the automark-actions variable.